Delete button added to Brands list
Additional button removed from some pages and new Delete button added to Brands list
