+++ public/css/app.css
@@ -1852,9 +1852,6 @@ html,  body,  #app{
 .whitespace-nowrap{
   white-space: nowrap;
 }
-.break-all{
-  word-break: break-all;
-}
 .rounded-md{
   border-radius: 0.375rem;
 }
@@ -2715,22 +2712,22 @@ html,  body,  #app{
   --tw-shadow-colored: 0 10px 15px -3px var(--tw-shadow-color), 0 4px 6px -4px var(--tw-shadow-color);
   box-shadow: var(--tw-ring-offset-shadow, 0 0 #0000), var(--tw-ring-shadow, 0 0 #0000), var(--tw-shadow);
 }
+.disabled\:bg-warm-blue:disabled{
+  --tw-bg-opacity: 1;
+  background-color: rgb(16 122 216 / var(--tw-bg-opacity));
+}
 .disabled\:bg-black:disabled{
   --tw-bg-opacity: 1;
   background-color: rgb(21 22 48 / var(--tw-bg-opacity));
 }
-.disabled\:bg-warm-blue:disabled{
-  --tw-bg-opacity: 1;
-  background-color: rgb(16 122 216 / var(--tw-bg-opacity));
+.disabled\:bg-opacity-75:disabled{
+  --tw-bg-opacity: 0.75;
 }
 .disabled\:bg-opacity-20:disabled{
   --tw-bg-opacity: 0.2;
 }
 .disabled\:bg-opacity-10:disabled{
   --tw-bg-opacity: 0.1;
-}
-.disabled\:bg-opacity-75:disabled{
-  --tw-bg-opacity: 0.75;
 }
 .disabled\:text-black:disabled{
   --tw-text-opacity: 1;
