+++ public/js/node_modules_moment_locale_sync_recursive_-resources_js_Pages_Customers_CustomersView_vue.js
@@ -179,8 +179,7 @@ __webpack_require__.r(__webpack_exports__);
 /* harmony import */ var vuex__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! vuex */ "./node_modules/vuex/dist/vuex.esm-bundler.js");
 /* harmony import */ var _heroicons_vue_24_solid__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! @heroicons/vue/24/solid */ "./node_modules/@heroicons/vue/24/solid/esm/PencilIcon.js");
 /* harmony import */ var _heroicons_vue_24_solid__WEBPACK_IMPORTED_MODULE_8__ = __webpack_require__(/*! @heroicons/vue/24/solid */ "./node_modules/@heroicons/vue/24/solid/esm/ArchiveBoxXMarkIcon.js");
-/* harmony import */ var _heroicons_vue_24_solid__WEBPACK_IMPORTED_MODULE_9__ = __webpack_require__(/*! @heroicons/vue/24/solid */ "./node_modules/@heroicons/vue/24/solid/esm/ArrowDownTrayIcon.js");
-/* harmony import */ var _heroicons_vue_24_outline__WEBPACK_IMPORTED_MODULE_10__ = __webpack_require__(/*! @heroicons/vue/24/outline */ "./node_modules/@heroicons/vue/24/outline/esm/FunnelIcon.js");
+/* harmony import */ var _heroicons_vue_24_outline__WEBPACK_IMPORTED_MODULE_9__ = __webpack_require__(/*! @heroicons/vue/24/outline */ "./node_modules/@heroicons/vue/24/outline/esm/FunnelIcon.js");
 /* harmony import */ var _Components_Elements_Headings_SidebarH2__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../Components/Elements/Headings/SidebarH2 */ "./resources/js/Components/Elements/Headings/SidebarH2.vue");
 /* harmony import */ var _Partials_NamingFilter__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../Partials/NamingFilter */ "./resources/js/Partials/NamingFilter.vue");
 /* harmony import */ var _Partials_Pagination__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ../../Partials/Pagination */ "./resources/js/Partials/Pagination.vue");
@@ -240,8 +239,7 @@ __webpack_require__.r(__webpack_exports__);
       useStore: vuex__WEBPACK_IMPORTED_MODULE_5__.useStore,
       Edit: _heroicons_vue_24_solid__WEBPACK_IMPORTED_MODULE_7__["default"],
       Delete: _heroicons_vue_24_solid__WEBPACK_IMPORTED_MODULE_8__["default"],
-      Download: _heroicons_vue_24_solid__WEBPACK_IMPORTED_MODULE_9__["default"],
-      Filter: _heroicons_vue_24_outline__WEBPACK_IMPORTED_MODULE_10__["default"],
+      Filter: _heroicons_vue_24_outline__WEBPACK_IMPORTED_MODULE_9__["default"],
       Heading2: _Components_Elements_Headings_SidebarH2__WEBPACK_IMPORTED_MODULE_1__["default"],
       CustomersFilter: _Partials_NamingFilter__WEBPACK_IMPORTED_MODULE_2__["default"],
       Pagination: _Partials_Pagination__WEBPACK_IMPORTED_MODULE_3__["default"],
@@ -936,13 +934,9 @@ var _hoisted_14 = {
 };
 var _hoisted_15 = ["href"];
 var _hoisted_16 = {
-  key: 1,
-  "class": "basis-1 flex justify-center items-center w-6 h-6 min-w-[24px] rounded bg-green font-medium text-blue"
-};
-var _hoisted_17 = {
   "class": "w-full flex justify-center items-center"
 };
-var _hoisted_18 = {
+var _hoisted_17 = {
   key: 1,
   "class": "w-full block relative text-center text-xl opacity-80 text-black font-semibold tracking-wider m-auto"
 };
@@ -1022,17 +1016,15 @@ function render(_ctx, _cache, $props, $setup, $data, $options) {
           "class": "w-4 h-4 text-white"
         })], 8
         /* PROPS */
-        , _hoisted_15)) : (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)("v-if", true), _ctx.$page.props.abilities.canEditCustomers ? ((0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)("button", _hoisted_16, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)($setup["Download"], {
-          "class": "w-4 h-4 text-white"
-        })])) : (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)("v-if", true)])) : (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)("v-if", true)]);
+        , _hoisted_15)) : (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)("v-if", true)])) : (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)("v-if", true)]);
       }), 128
       /* KEYED_FRAGMENT */
-      ))])])]), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)(" Pagination section "), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_17, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)($setup["Pagination"], {
+      ))])])]), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)(" Pagination section "), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_16, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)($setup["Pagination"], {
         links: $props.customers.links,
         params: $props.params
       }, null, 8
       /* PROPS */
-      , ["links", "params"])]), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)(" Empty customers list message "), Object.keys($props.customers.data).length <= 0 ? ((0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)("strong", _hoisted_18, " No customers found ")) : (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)("v-if", true), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)(" Delete Modal Component "), $setup.store.getters.deleteModal ? ((0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createBlock)($setup["DeleteModal"], {
+      , ["links", "params"])]), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)(" Empty customers list message "), Object.keys($props.customers.data).length <= 0 ? ((0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)("strong", _hoisted_17, " No customers found ")) : (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)("v-if", true), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)(" Delete Modal Component "), $setup.store.getters.deleteModal ? ((0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createBlock)($setup["DeleteModal"], {
         key: 2
       })) : (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)("v-if", true)])];
     }),
