+++ public/js/node_modules_moment_locale_sync_recursive_-resources_js_Pages_Invoices_ViewInvoices_vue.js
@@ -878,14 +878,9 @@ var _hoisted_26 = {
 };
 var _hoisted_27 = ["href"];
 var _hoisted_28 = {
-  key: 1,
-  href: "#",
-  "class": "basis-1 flex justify-center items-center w-6 h-6 min-w-[24px] rounded bg-green font-medium text-blue"
-};
-var _hoisted_29 = {
   "class": "w-full flex justify-center items-center"
 };
-var _hoisted_30 = {
+var _hoisted_29 = {
   key: 1,
   "class": "w-full px-4 py-12 text-center text-lg font-semibold text-black text-opacity-75 tracking-wider"
 };
@@ -950,17 +945,15 @@ function render(_ctx, _cache, $props, $setup, $data, $options) {
           "class": "w-4 h-4 text-white"
         })], 8
         /* PROPS */
-        , _hoisted_27)) : (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)("v-if", true), _ctx.$page.props.abilities.canExportInvoices ? ((0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)("a", _hoisted_28, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)($setup["Download"], {
-          "class": "w-4 h-4 text-white"
-        })])) : (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)("v-if", true)])) : (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)("v-if", true)]);
+        , _hoisted_27)) : (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)("v-if", true)])) : (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)("v-if", true)]);
       }), 128
       /* KEYED_FRAGMENT */
-      ))])])]), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)(" Pagination section "), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_29, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)($setup["Pagination"], {
+      ))])])]), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)(" Pagination section "), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("div", _hoisted_28, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)($setup["Pagination"], {
         links: $props.invoices.links,
         params: $props.params
       }, null, 8
       /* PROPS */
-      , ["links", "params"])])])) : ((0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)("h2", _hoisted_30, " No invoice found! "))];
+      , ["links", "params"])])])) : ((0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)("h2", _hoisted_29, " No invoice found! "))];
     }),
     _: 1
     /* STABLE */
